saves author but not answer content

+++ app/routes/question.js
@@ -15,9 +15,27 @@ export default Ember.Route.extend({
       this.transitionTo('index');
     },
 
-    destroyQuestion(question) {
-      question.destroyRecord();
-      this.transitionTo('index');
+    destroyQuestion(question){
+      if(confirm('Delete this question?')){
+        var answer_deletions = question.get('answer').map(function(answer){
+          return answer.destroyRecord();
+        });
+        Ember.RSVP.all(answer_deletions).then(function(){
+          return question.destroyRecord();
+        });
+        this.transitionTo('index');
+      }
+    },
+
+    saveAnswer3(params){
+      var newAnswer = this.store.createRecord('answer', params);
+      var question = params.question;
+
+      question.get('answer').addObject(newAnswer);
+      newAnswer.save().then(function(){
+        return question.save();
+      });
+      this.transitionTo('question', params.question);
     }
   }
 });
